pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"
requires_python = ">=3.10"



[project]
name = "moralis-streams-client"
authors = [{name = "Matt Krueger", email = "mkrueger@rstms.net"}]
readme = {file = "README.md", content-type = "text/markdown"}
license = {file = "LICENSE"}
keywords = ["moralis_streams_client"]
classifiers = [
  "Intended Audience :: Developers",

  "Natural Language :: English", 
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10"
]
dynamic = ["version", "description"] 

dependencies = [
  "click",
  "httpx",
  "attrs"
]

[tool.flit.module]
name = "moralis_streams_client"

[project.optional-dependencies]
dev = [
  "black",
  "bump2version",
  "coverage",
  "isort",
  "flake8",
  "flake8-length",
  "pytest",
  "python-box",
  "tox"
]
docs = [
  "m2r2",
  "sphinx",
  "sphinx-click",
  "sphinx-rtd-theme"
]

[project.urls]
Home = "https://github.com/rstms/moralis_streams_client"
[tool.black]
line-length = 79

[tool.isort]
profile = "black"
[tool.pytest]
collect_ignore = ['setup.py']


[project.scripts]
msc = "moralis_streams_client:msc"
